+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.3.3] - 2026-01-11 - FFXIV Recorder
+### Changed
+- Combat start now based on engage timer event as even ff act timestamp are not the same for everyone (should be close enough, can't be more in sync with the date we have here)
+
 ## [8.3.2] - 2026-01-11 - FFXIV Recorder
 ### Fixed
 - Added a very small wait time when end of combat is detected to be able to receive boss death info (ff combat end trigger is too fast)